moving files to various packages
