+++ src/main/scala/nl/knaw/dans/easy/pid/generator/PidGenerator.scala
@@ -13,12 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid
+package nl.knaw.dans.easy.pid.generator
 
 import java.io.File
 import java.lang.Math.pow
 
 import nl.knaw.dans.easy.pid.microservice.{DOI, PidType, Settings, URN}
+import nl.knaw.dans.easy.pid.DbBasedSeedStorage
 
 import scala.util.Try
