+++ src/main/scala/nl/knaw/dans/easy/pid/generator/SeedStorage.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid
+package nl.knaw.dans.easy.pid.generator
 
 import java.io.File
 import javax.persistence.Entity
