+++ src/main/scala/nl/knaw/dans/easy/pid/rest/RestService.scala
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.pid
+package nl.knaw.dans.easy.pid.rest
 
 import com.typesafe.config.Config
+import nl.knaw.dans.easy.pid.Service
 import org.eclipse.jetty.server.Server
 import org.eclipse.jetty.servlet.ServletContextHandler
 import org.scalatra.servlet.ScalatraListener
